Module Tjr_fs_shared.Buffers_from_btree

Buffer operations including blit, moved from Tjr_btree.

Don't use this module directly; instead use buf_ops#string_ etc

NOTE this uses value passing; to check linear usage, use the "safe_" version which tracks liveness of each buffer

type 'buf buf_ops = {

buf_create : int -> 'buf;

buf_length : 'buf -> int;

buf_to_string : src:'buf -> off:Int_like.offset -> len:Int_like.len -> string;

to_string : 'buf -> string;

of_string : string -> 'buf;

of_bytes : bytes -> 'buf;

blit : src:'buf -> src_off:Int_like.offset -> src_len:Int_like.len -> dst:'buf -> dst_off:Int_like.offset -> 'buf;

blit_bytes_to_buf : src:bytes -> src_off:Int_like.offset -> src_len:Int_like.len -> dst:'buf -> dst_off:Int_like.offset -> 'buf;

blit_string_to_buf : src:string -> src_off:Int_like.offset -> src_len:Int_like.len -> dst:'buf -> dst_off:Int_like.offset -> 'buf;

}

"mutable", fixed size buffers; it is an error to attempt to write beyond the end; prefer a functional repr. for now

$(FIXME("change default buffer implementation to bytes or bigstring for production"))

module B_ = Stdlib.BytesLabels

module Check_blit : functor (S : sig ... end) -> sig ... end

module String_ : sig ... end

module Abstract_string : sig ... end

Abstract buffer type, with underlying string representation; FIXME move to mutable implementation for production

module Bytes_ : sig ... end

module Abstract_bytes : sig ... end

Abstract buffer type, with underlying bytes representation

module Safe_bytes : sig ... end

module Bigstring_ : sig ... end

NOTE this does not check linearity; use something else for testing

val buf_ops : < abstract_bytes : Abstract_bytes.t buf_ops; abstract_string : Abstract_string.t buf_ops; ba : Bigstring.t buf_ops; bytes : bytes buf_ops; safe_bytes : Safe_bytes.t buf_ops; string : string buf_ops; >

Export all implementations under object value buf_ops