Module Tjr_fs_shared.File_ops

To make the code more uniform, we assume various "normal" file operations (ie over the local filesystem) are available in the monad. Of course, these operations are imperative since they use a real underlying filesystem.

type ('fd, 'blk, 't) file_ops = {

read_file : Tjr_lib.Tjr_file.Filenames.fn -> (string, 't) Tjr_monad.m;

write_string_to_file : fn:Tjr_lib.Tjr_file.Filenames.fn -> string -> (unit, 't) Tjr_monad.m;

stat : Tjr_lib.Tjr_file.Filenames.fn -> (Tjr_lib.Tjr_file.fds_t option, 't) Tjr_monad.m;

fd_from_file : fn:Tjr_lib.Tjr_file.Filenames.fn -> create:bool -> init:bool -> ('fd, 't) Tjr_monad.m;

close : 'fd -> (unit, 't) Tjr_monad.m;

blk_sz : int;

read_blk : 'fd -> int -> ('blk, 't) Tjr_monad.m;

write_blk : 'fd -> int -> 'blk -> (unit, 't) Tjr_monad.m;

}

This is minimal, so we can support blk-dev-on-fd initialization and finalization. pread and pwrite always use blk-sized amounts of data. read_blk and write_blk take a blk_index

module Lwt_file_ops : sig ... end

val lwt_file_ops : (Lwt_file_ops.fd, Lwt_file_ops.blk, Tjr_monad.lwt) file_ops