Module Tjr_fs_shared.Marshal_factory

type ('a, 'buf) mshlr = {

max_elt_sz : int;

mshl : 'a -> ('buf * int) -> 'buf * int;

umshl : 'buf -> int -> 'a * int;

}

Generic marshaller type, assuming max_elt_sz is known (alternative is to marshal and check for end of buffer explicitly)

type ('k, 'v, 'buf) kv_mshlr = {

k_mshlr : ('k, 'buf) mshlr;

v_mshlr : ('v, 'buf) mshlr;

}

A pair of marshallers, one for keys, one for values

module Make_marshaller : functor (X : sig ... end) -> sig ... end

Functor to construct a marshaller from a type deriving bin_io

type arg =

| A1_int_option__ba_buf

| A2_blk_id_opt__ba_buf

| A3_int__ba_buf

| A4_str_256__ba_buf

| A5_int_int_kv_mshlr

| A6_s256_int_kv_mshlr

| A7_s256_s256_kv_mshlr

type res =

| R1 of (int option, Tjr_fs_shared__.Buf_ops.ba_buf) mshlr

| R2 of (Tjr_fs_shared__.Blk_intf.Blk_id_as_int.blk_id option, Tjr_fs_shared__.Buf_ops.ba_buf) mshlr

| R3 of (int, Tjr_fs_shared__.Buf_ops.ba_buf) mshlr

| R4 of (Str_256.str_256, Tjr_fs_shared__.Buf_ops.ba_buf) mshlr

| R5 of (int, int, Tjr_fs_shared__.Buf_ops.ba_buf) kv_mshlr

| R6 of (Str_256.str_256, int, Tjr_fs_shared__.Buf_ops.ba_buf) kv_mshlr

| R7 of (Str_256.str_256, Str_256.str_256, Tjr_fs_shared__.Buf_ops.ba_buf) kv_mshlr

val make_1 : (int option, Bin_prot.Common.buf) mshlr

val make_2 : (Tjr_fs_shared__.Blk_intf.Blk_id_as_int.blk_id option, Bin_prot.Common.buf) mshlr

val make_3 : (int, Bin_prot.Common.buf) mshlr

val int_mshlr : (int, Bin_prot.Common.buf) mshlr

val make_4 : (Str_256.str_256, Bin_prot.Common.buf) mshlr

val str_256_mshlr : (Str_256.str_256, Bin_prot.Common.buf) mshlr

val make_5 : (int, int, Bin_prot.Common.buf) kv_mshlr

val make_6 : (Str_256.str_256, int, Bin_prot.Common.buf) kv_mshlr

val make_7 : (Str_256.str_256, Str_256.str_256, Bin_prot.Common.buf) kv_mshlr