Module Blk_intf.Blk_ops

type 'blk blk_ops = {

blk_sz : blk_sz;

of_string : string -> 'blk;

to_string : 'blk -> string;

of_bytes : bytes -> 'blk;

to_bytes : 'blk -> bytes;

}

NOTE Conversion to a blk is expected (and the common instances do, indeed!) to pad if the string/bytes are not long enough.

NOTE since blk_sz is expected to be fixed for a given blk type, we don't include as a parameter on the main interface methods