Module Tjr_fs_shared__.Shared_map_ops

module Fid_map_ops : sig ... end

Simple map operations: find, insert and delete (no insert_many).

module Map_ops_type : sig ... end

include Map_ops_type

Default map operations, with insert_many

type ('k, 'v, 't) map_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

insert_many : 'k -> 'v -> ('k * 'v) list -> (('k * 'v) list, 't) Tjr_monad.m;

}

Map operations: find, insert, insert_many and delete.

insert_many attempts to insert as many as possible in a single operation, and returns the remainder, and so is typically called in a loop (see Map_ops).

The interfaces are heavily parameterized. To understand the interfaces, we need to introduce the following:

Keys, represented by type variable 'k

Values, by type var 'v

Page/block references, 'r; vars blk_id

Phantom monad type, 't

The operations execute in the monad.

module Internal_make_insert_many_insert_all : sig ... end