Module Tjr_fs_shared__Blk_intf

Blk-related interfaces

The main choices:

blk implementation (string, bytes, buf/bigarray etc)

blk_id (presumably int, but perhaps a pair of (dev,int) )

blk_dev (in mem, on file, on dev etc)

module Blk_sz : sig ... end

type blk_sz = Blk_sz.blk_sz

val blk_sz_4096 : Blk_sz.blk_sz

module Blk_id_as_int : sig ... end

This is a common instance of blk_id; we don't open it by default because we want most code to be independent of the exact repn. of blk_id NOTE do not open this module FIXME replace with { blk_id:'a} and { blk_id:int} record

module Blk_ops : sig ... end

include Blk_ops

type 'blk blk_ops = {

blk_sz : blk_sz;

of_string : string -> 'blk;

to_string : 'blk -> string;

of_bytes : bytes -> 'blk;

to_bytes : 'blk -> bytes;

}

NOTE Conversion to a blk is expected (and the common instances do, indeed!) to pad if the string/bytes are not long enough.

NOTE since blk_sz is expected to be fixed for a given blk type, we don't include as a parameter on the main interface methods

module Blk_dev_ops : sig ... end

include Blk_dev_ops

type ('blk_id, 'blk, 't) blk_dev_ops = {

blk_sz : blk_sz;

write : blk_id:'blk_id -> blk:'blk -> (unit, 't) Tjr_monad.m;

read : blk_id:'blk_id -> ('blk, 't) Tjr_monad.m;

write_many : ('blk_id * 'blk) list -> (unit, 't) Tjr_monad.m;

}

A block device: read and write blocks.

module Blk_allocator_ops : sig ... end

include Blk_allocator_ops

type ('blk_id, 't) blk_allocator_ops = {

blk_alloc : unit -> ('blk_id, 't) Tjr_monad.m;

blk_free : 'blk_id -> (unit, 't) Tjr_monad.m;

}

NOTE we assume alloc never fails, or that error is handled elsewhere in the monad; fields were named alloc and free