Module Tjr_fs_shared__Configs

Utilities for handling config files.

The environment variable "TJR_CONFIGS_DIR", if set, contains the config files.

At some point after module initialization, the "run_load_hooks" method is called (once only). This is responsible for running all registered hooks.

Each individual library typically registers a hook with the "add_load_hook" method. This function will usually check the configs_dir, and if not set it will try to load a config from the current directory.

At the moment, if TJR_CONFIGS_DIR is not set, we set it anyway to /tmp, to avoid libraries trying to load configs from the current directory. So best practice is to always have this var set to point to your configs directory.

val add_load_hook : (unit -> unit) -> unit

val run_load_hooks : unit -> unit

val configs_dir : unit -> string