Module Tjr_fs_shared

Common types and defns for file system work.

A collection of the main types provided by this library.

A record for the pair of an initial state and related operations

type ('a, 'b) initial_state_and_ops = ('a, 'b) Tjr_fs_shared__.Internal_intf.initial_state_and_ops = {

initial_state : 'a;

ops : 'b;

}

Int-like types

module Int_like : sig ... end

Int-like types with unboxed record wrappers, for documentation purposes

Buffers

include Tjr_fs_shared__.Buf_ops

type 'buf buf_ops = {

create : int -> 'buf;
Assumed to be zero-ed

get : int -> 'buf -> char;

set : int -> char -> 'buf -> 'buf;

len : 'buf -> int;

}

val chr0 : char

type ba_buf = (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type ba_buf_ops = ba_buf buf_ops

type by_buf_ops = bytes buf_ops

module Buf_factory : sig ... end

Common buffer operations

module Buf_as_bigarray = Buf_factory.Buf_as_bigarray

module Buf_as_bytes = Buf_factory.Buf_as_bytes

val ba_buf_ops : Buf_as_bigarray.ba_buf_ops

val by_buf_ops : Buf_as_bytes.by_buf_ops

module Buffers_from_btree : sig ... end

FIXME move this elsewhere eg fs_shared (and combine with other similar modules)

Block-related types

include Tjr_fs_shared__.Blk_intf

module Blk_sz = Tjr_fs_shared__.Blk_intf.Blk_sz

type blk_sz = Blk_sz.blk_sz

val blk_sz_4096 : Blk_sz.blk_sz

module Blk_id_as_int = Tjr_fs_shared__.Blk_intf.Blk_id_as_int

This is a common instance of blk_id; we don't open it by default because we want most code to be independent of the exact repn. of blk_id NOTE do not open this module FIXME replace with { blk_id:'a} and { blk_id:int} record

module Blk_ops = Tjr_fs_shared__.Blk_intf.Blk_ops

include Blk_ops

type 'blk blk_ops = {

blk_sz : blk_sz;

of_string : string -> 'blk;

to_string : 'blk -> string;

of_bytes : bytes -> 'blk;

to_bytes : 'blk -> bytes;

}

NOTE Conversion to a blk is expected (and the common instances do, indeed!) to pad if the string/bytes are not long enough.

NOTE since blk_sz is expected to be fixed for a given blk type, we don't include as a parameter on the main interface methods

module Blk_dev_ops = Tjr_fs_shared__.Blk_intf.Blk_dev_ops

include Blk_dev_ops

type ('blk_id, 'blk, 't) blk_dev_ops = {

blk_sz : blk_sz;

write : blk_id:'blk_id -> blk:'blk -> (unit, 't) Tjr_monad.m;

read : blk_id:'blk_id -> ('blk, 't) Tjr_monad.m;

write_many : ('blk_id * 'blk) list -> (unit, 't) Tjr_monad.m;

}

A block device: read and write blocks.

module Blk_allocator_ops = Tjr_fs_shared__.Blk_intf.Blk_allocator_ops

include Blk_allocator_ops

type ('blk_id, 't) blk_allocator_ops = {

blk_alloc : unit -> ('blk_id, 't) Tjr_monad.m;

blk_free : 'blk_id -> (unit, 't) Tjr_monad.m;

}

NOTE we assume alloc never fails, or that error is handled elsewhere in the monad; fields were named alloc and free

Block-related implementations

module Blk_factory : sig ... end

module Blk_dev_factory : sig ... end

Common blk devs; don't open

module Root_block : sig ... end

Kv ops

module Kvop : sig ... end

A concrete type for insert and delete operations. Don't open.

type ('k, 'v) kvop = ('k, 'v) Kvop.kvop

module Kvop_map = Kvop.Kvop_map

(Shared) Map ops

module Shared_map_ops : sig ... end

Map operations find,ins,del, in; insert_all; make_insert_many FIXME? note may clash with other "map_ops" so we don't include at the top-level

Small strings, leq 256 bytes

module Str_256 : sig ... end

type str_256 = Str_256.str_256

Maps with key traversal: get_next_binding, get_prev_binding

module Map_with_key_traversal : sig ... end

Write back cache

module Write_back_cache : sig ... end

A wrapper around pqwy/lru.

module Write_back_cache_v2_DONT_USE = Tjr_fs_shared__.Write_back_cache_v2

File operations

module File_ops : sig ... end

To make the code more uniform, we assume various "normal" file operations (ie over the local filesystem) are available in the monad. Of course, these operations are imperative since they use a real underlying filesystem.

val lwt_file_ops : (File_ops.Lwt_file_ops__ba_buf.fd, File_ops.Lwt_file_ops__ba_buf.blk, Tjr_monad.lwt) File_ops.file_ops

Marshalling

module Marshal_factory : sig ... end

type ('a, 'buf) mshlr = ('a, 'buf) Marshal_factory.mshlr = {

max_elt_sz : int;

mshl : 'a -> ('buf * int) -> 'buf * int;

umshl : 'buf -> int -> 'a * int;

}

type ('k, 'v, 'buf) kv_mshlr = ('k, 'v, 'buf) Marshal_factory.kv_mshlr = {

k_mshlr : ('k, 'buf) mshlr;

v_mshlr : ('v, 'buf) mshlr;

}

Std types

module Std_types : sig ... end

Testing

module Test : sig ... end

Support for basic testing, controlled by optcomp

Log

module Log : sig ... end

Runtime config support

val runtime_config_factory : < make : default:'a option -> filename:string -> 'a Tjr_fs_shared__.Runtime_config_factory.runtime_config -> 'a Stdlib.Lazy.t; >